Simplify int to uint64 conversion to address CI failure

+++ utils.go
@@ -954,7 +954,7 @@ func copySortedFunc[E any](slice []E, less func(a, b E) bool) []E {
 	return sliceCopy
 }
 
-func contains(s []int, x int) bool {
+func contains(s []uint64, x uint64) bool {
 	for _, v := range s {
 		if v == x {
 			return true
